+++ src/main/kotlin/com/shinriyo/slangmate/RunSlangAction.kt
@@ -7,15 +7,22 @@ import com.intellij.openapi.ui.Messages
 import java.io.BufferedReader
 import java.io.InputStreamReader
 
-class RunSlangAction : AnAction("Run Slang") {
+class RunSlangAction : AnAction() {
+    init {
+        templatePresentation.apply {
+            text = "Run Slang"
+            description = "Execute slang command"
+        }
+    }
+    
     override fun actionPerformed(e: AnActionEvent) {
-        val project: Project? = e.project ?: return
+        val project: Project = e.project ?: return
 
         val command = listOf("fvm", "flutter", "pub", "run", "slang")
 
         try {
             val processBuilder = ProcessBuilder(command)
-                .directory(project?.basePath?.let { java.io.File(it) }) // プロジェクトのルートで実行
+                .directory(project.basePath?.let { java.io.File(it) }) // プロジェクトのルートで実行
                 .redirectErrorStream(true) // 標準エラーも含める
 
             val process = processBuilder.start()